+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
   new value, suitable for serialization to a JSON string.  Easily generate the
   payload to send in an HTTP PUT to a web service.
 
+- ***New attribute type: json***
+
 ## 1.3.0
 
 - **Breaking change**: Parsing an integer to a time attribute, the integer is